Increased the delay after a rest.
The RFXtrx wasn't reliably responsing to a status update after 0.1
seconds. It *should* work after 0.05 seconds, however, after some
testing it seems to work only reliably after 0.4 seconds. This
change increases the delay, but it seems like we could do with a
better approach for this. Something that keeps writing until we
get a response?

This fix builds on the work in #15